requirements.txt
uvicorn~=0.22.0
fastapi~=0.95.1
orhelper~=0.1.3
numpy~=1.24.3
Pillow~=9.5.0
matplotlib~=3.7.1
scipy~=1.10.1
deap~=1.3.3